+++ setup.py
@@ -1,11 +1,11 @@
 from distutils.core import setup
 
 setup(name='number-detector',
-      version=open("object_detector/_version.py").readlines()[-1].split()[-1].strip("\"'"),
+      version=open("digit_detector/_version.py").readlines()[-1].split()[-1].strip("\"'"),
       description='SVHN number detector and recognizer',
       author='jeongjoonsup',
       author_email='penny4860@gmail.com',
       url='https://penny4860.github.io/',
-      packages=['object_detector'],  
+      packages=['digit_detector'],  
      )
